+++ .eslintrc.json
@@ -18,6 +18,7 @@
     }
   },
   "rules": {
-    "no-unused-vars": "warn"
+    "no-unused-vars": "warn",
+    "react/prop-types": 0
   }
 }
