+++ src/redux/reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux'
+import authReducer from './auth_reducer';
+
+/**
+ * Connect the reducers here
+ */
+
+export default combineReducers({
+  Auth: authReducer,
+});
